refactor(scripts): move support to separate dir
